# openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Todo API with Supabase Authentication",
    "description": "API RESTful de tareas con autenticación Supabase, PostgreSQL y Redis",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Local server"
    }
  ],
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
      "AuthCredentials": {
        "type": "object",
        "required": ["email", "password"],
        "properties": {
          "email": {
            "type": "string",
            "format": "email",
            "example": "user@example.com"
          },
          "password": {
            "type": "string",
            "format": "password",
            "example": "Password123!"
          }
        }
      },
      "Task": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "example": 1
          },
          "title": {
            "type": "string",
            "example": "Comprar leche"
          },
          "done": {
            "type": "boolean",
            "example": false
          }
        }
      },
      "TaskInput": {
        "type": "object",
        "required": ["title"],
        "properties": {
          "title": {
            "type": "string",
            "example": "Nueva tarea"
          },
          "done": {
            "type": "boolean",
            "example": false
          }
        }
      },
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string",
            "example": "Invalid or expired token"
          }
        }
      }
    }
  },
  "paths": {
    "/": {
      "get": {
        "summary": "Health check y estado del motor de almacenamiento",
        "tags": ["System"],
        "responses": {
          "200": {
            "description": "API operational"
          }
        }
      }
    },
    "/auth/signup": {
      "post": {
        "summary": "Registrar un nuevo usuario",
        "tags": ["Authentication"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AuthCredentials"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Usuario creado exitosamente"
          },
          "400": {
            "description": "Datos de registro inválidos"
          }
        }
      }
    },
    "/auth/login": {
      "post": {
        "summary": "Iniciar sesión y obtener tokens JWT",
        "tags": ["Authentication"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AuthCredentials"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Autenticación exitosa con access_token"
          },
          "401": {
            "description": "Credenciales inválidas"
          }
        }
      }
    },
    "/auth/logout": {
      "post": {
        "summary": "Cerrar sesión (Endpoint Protegido)",
        "tags": ["Authentication"],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "responses": {
          "204": {
            "description": "Sesión cerrada exitosamente (Sin contenido)"
          },
          "401": {
            "description": "Token inválido o expirado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/public/info": {
      "get": {
        "summary": "Información pública de bienvenida",
        "tags": ["Public"],
        "responses": {
          "200": {
            "description": "Información pública devuelta"
          }
        }
      }
    },
    "/protected/profile": {
      "get": {
        "summary": "Perfil de usuario protegido",
        "tags": ["Protected"],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Perfil del usuario autenticado"
          },
          "401": {
            "description": "Token inválido o expirado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/protected/dashboard": {
      "get": {
        "summary": "Dashboard protegido",
        "tags": ["Protected"],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Dashboard del usuario autenticado"
          },
          "401": {
            "description": "Token inválido o expirado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/todos": {
      "get": {
        "summary": "Obtener todas las tareas",
        "tags": ["Tasks"],
        "responses": {
          "200": {
            "description": "Lista de tareas"
          }
        }
      },
      "post": {
        "summary": "Crear una nueva tarea",
        "tags": ["Tasks"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TaskInput"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Tarea creada"
          },
          "400": {
            "description": "Datos de entrada inválidos"
          }
        }
      }
    },
    "/api/todos/{id}": {
      "get": {
        "summary": "Obtener tarea por ID",
        "tags": ["Tasks"],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Tarea encontrada"
          },
          "404": {
            "description": "Tarea no encontrada"
          }
        }
      },
      "put": {
        "summary": "Actualizar tarea por ID",
        "tags": ["Tasks"],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TaskInput"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Tarea actualizada"
          },
          "404": {
            "description": "Tarea no encontrada"
          }
        }
      },
      "delete": {
        "summary": "Eliminar tarea por ID",
        "tags": ["Tasks"],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Tarea eliminada"
          },
          "404": {
            "description": "Tarea no encontrada"
          }
        }
      }
    }
  }
}